+++ src/signals/incident/components/form/MapSelectors/Asset/Selector/WfsLayer/ReportedLayer/index.ts
@@ -1,0 +1,3 @@
+// SPDX-License-Identifier: MPL-2.0
+// Copyright (C) 2021 Gemeente Amsterdam
+export { default } from './ReportedLayer'
